WIP: Use Wasm struct to metadata to pass as argument to the `j.l.Class`
The `j.l.Class` constructor receives a JS object containing metadata, and the methods in `j.l.Class` depend on that metadata.
This commit attempts to remove the dependency on the JS object for metadata by representing it using a WasmGC struct.

When calling methods (`ApplyStatically`) on `j.l.Class`, the function calls will be replaced by implementations that rely on the metadata represented in the WasmGC struct.
One issue is that we cannot replace the function call for `j.l.Class#cast` because it's inlined by the optimizer.

Do you think this is the right approach, and is there a workaround?

---

Alternative approaches

(1) Modify the `j.l.Class` method implementations to avoid accessing the JS object containing metadata,
and instead access the WasmGC struct. I'm unsure how to access the metadata at the library level; perhaps some new IR nodes are needed?

(2) Compile `JSSelect`/`JSMethodApply` on the `ScalaJSClassData` to field access on the WasmGC struct that contains the metadata.
I'm not sure how to determine if the qualifier of `JSSelect` is `ScalaJSClassData`, as `qualifier.tpe` on `JSSelect` returns `AnyType` in this case.
Also, I don't think this is the correct approach to removing JS interop. I believe the right way to avoid JS interop involving `js.Object` is to avoid accessing the `js.Object` altogether, and `JSSelect`/`JSMethodApply` shouldn’t appear at link-time under pure-WASM.